requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.5
djangorestframework==3.14.0
dnspython==2.4.2
eventlet==0.33.3
greenlet==2.0.2
kombu==5.3.2
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-magic==0.4.27
# python-magic-bin==0.4.14
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.8
